# Increase glTF parsing performance