First of all, great project!

While testing your usage of [qmuntal/gltf](https://github.com/qmuntal/gltf) I noticed `LoadGLTFData` consumes quite a lot of memory compared to the size of the parsed glTF file. For example, `./example/logo/tetra3d.glb` weights ~58KB and `LoadGLTFData` consumes ~3MB and produces ~27k allocations!

My take is that most of the memory consumed by that function is accidental. A sane size to in-memory ratio would be 2 or 3x, not 58x.

This PR contains several memory improvements, which as side effect also have some net speed gains:

```txt
C:\Users\qmuntaldiaz\code\Tetra3d>benchstat old.txt new.txt
name             old time/op    new time/op    delta
LoadGLTFData-12    1.97ms ±10%    1.36ms ±15%  -30.67%  (p=0.000 n=9+9)

name             old speed      new speed      delta
LoadGLTFData-12  29.9MB/s ± 9%  43.1MB/s ±13%  +44.27%  (p=0.000 n=9+9)

name             old B/op       new B/op       delta
LoadGLTFData-12    3.17MB ± 0%    1.64MB ± 0%  -48.19%  (p=0.000 n=10+10)

name             old allocs/op  new allocs/op  delta
LoadGLTFData-12     27.6k ± 0%     25.5k ± 0%   -7.60%  (p=0.000 n=10+9)
```