@SolarLune 2022-03-09
Oh, wow, thanks a lot for the help - this much of a massive memory savings is great to see. I've reviewed the PR and made a comment - thanks a lot, again for submitting it!